Fix sign-compare warning
Fix warning comparing long int to long unsigned int.

+++ src/output_docbook.cpp
@@ -754,7 +754,7 @@ DocBookDocumentSectioning hasDocumentSectioning(ParagraphList const &paragraphs,
 	bool documentHasSections = false;
 
 	while (bpit < epit) {
-		LASSERT(bpit < paragraphs.size(), return make_tuple(documentHasSections, bpit));
+		LASSERT(static_cast<size_t>(bpit) < paragraphs.size(), return make_tuple(documentHasSections, bpit));
 
 		Layout const &style = paragraphs[bpit].layout();
 		documentHasSections |= isLayoutSectioningOrSimilar(style);
